centralise toolkit icon generator
