+++ gui/utils/utils.js
@@ -226,3 +226,24 @@ export const createInternalId = () => {
 
   return newId;
 }
+
+export const returnToolkitIcon = (toolkitName) => {
+  const toolkitData = [
+    { name: 'Jira Toolkit', imageSrc: '/images/jira_icon.svg' },
+    { name: 'Email Toolkit', imageSrc: '/images/gmail_icon.svg' },
+    { name: 'Google Calendar Toolkit', imageSrc: '/images/google_calender_icon.svg' },
+    { name: 'GitHub Toolkit', imageSrc: '/images/github_icon.svg' },
+    { name: 'Google Search Toolkit', imageSrc: '/images/google_search_icon.svg' },
+    { name: 'Searx Toolkit', imageSrc: '/images/searx_icon.svg' },
+    { name: 'Slack Toolkit', imageSrc: '/images/slack_icon.svg' },
+    { name: 'Web Scraper Toolkit', imageSrc: '/images/webscraper_icon.svg' },
+    { name: 'Twitter Toolkit', imageSrc: '/images/twitter_icon.svg' },
+    { name: 'Google SERP Toolkit', imageSrc: '/images/google_serp_icon.svg' },
+    { name: 'File Toolkit', imageSrc: '/images/filemanager_icon.svg' },
+    { name: 'CodingToolkit', imageSrc: '/images/app-logo-light.png' },
+    { name: 'Image Generation Toolkit', imageSrc: '/images/app-logo-light.png' },
+  ];
+
+  const toolkit = toolkitData.find((tool) => tool.name === toolkitName);
+  return toolkit ? toolkit.imageSrc : '/images/custom_tool.svg';
+}
